+++ tests/audio/cd_audio_runtime_tests.c
@@ -35,7 +35,7 @@ void StepCdTrackRequest(void) { s_trackSteps++; }
 void StepCdPlayRequest(void) { s_playSteps++; }
 void StepCdPauseRequest(void) { s_pauseSteps++; }
 void StepCdVolumeFade(void) { s_fadeSteps++; }
-long HostCdAudioEnded(void) { return s_hostEnded; }
+s32 HostCdAudioEnded(void) { return s_hostEnded; }
 void BuildCdTrackTable(void) { s_buildCalls++; }
 void SetCdVolume(s32 volume) { s_setVolume = volume; }
